[refactor] remove httparty dependency
HTTParty is the only gem dependency of the bot.
I thought it was redundant to maintain a single
dependency especially when it was achievable with
the standard net/http library except that it would
some extra lines of code.
I thought the extra lines of code was worth it
rather than having to maintain a single dependency.